@@ -82,11 +82,11 @@
     "moment": "^2.29.1",
     "node-sass": "4.14.1",
     "prop-types": "^15.7.2",
-    "react": "^17.0.2",
-    "react-bootstrap": "1.3.0",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-bootstrap": "2.4.0",
+    "react-dom": "^18.1.0",
     "react-icons": "^4.3.1",
-    "react-router-dom": "5.2.0",
+    "react-router-dom": "6.3.0",
     "react-scripts": "5.0.0",
     "uuid": "^8.3.2",
     "web-vitals": "^2.1.2"
