@@ -1,7 +1,7 @@
 import React from 'react';
 import Signup from './pages/Signup';
 import { AuthProvider } from './context/AuthContext';
-import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Dashboard from './pages/Dashboard';
 import Login from './pages/Login';
 import PrivateRoute from './pages/PrivateRoute';
@@ -12,27 +12,44 @@ import { ProjectsPage } from './pages/ProjectsPage';
 import { ROUTES } from './config/constants';
 import 'bootstrap/dist/css/bootstrap.css';
 
-import { createBrowserHistory } from 'history';
-
-const history = createBrowserHistory();
-
 export const App = () => (
   <div>
-    <Router history={history}>
+    <Router>
       <AuthProvider>
-        <Switch>
-          <PrivateRoute exact path={ROUTES.DASHBOARD} component={Dashboard} />
-          <PrivateRoute exact path={ROUTES.HOME} component={ProjectsPage} />
-          <PrivateRoute
-            exact
-            path={ROUTES.UPDATE_PROFILE}
-            component={UpdateProfile}
+        <Routes>
+          <Route
+            path={ROUTES.DASHBOARD}
+            element={
+              <PrivateRoute>
+                {' '}
+                <Dashboard />
+              </PrivateRoute>
+            }
           />
-          <Route path={ROUTES.SIGN_UP} component={Signup} />
-          <Route path={ROUTES.LOGIN} component={Login} />
-          <Route path={ROUTES.FORGOT_PASSWORD} component={ForgotPassword} />
-          <Route path="*" component={NotFound} />
-        </Switch>
+          <Route
+            path={ROUTES.HOME}
+            element={
+              <PrivateRoute>
+                {' '}
+                <ProjectsPage />
+              </PrivateRoute>
+            }
+          />
+          <Route
+            path={ROUTES.UPDATE_PROFILE}
+            element={
+              <PrivateRoute>
+                {' '}
+                <UpdateProfile />
+              </PrivateRoute>
+            }
+          />
+
+          <Route path={ROUTES.SIGN_UP} element={<Signup />} />
+          <Route path={ROUTES.LOGIN} element={<Login />} />
+          <Route path={ROUTES.FORGOT_PASSWORD} element={<ForgotPassword />} />
+          <Route path="/*" element={<NotFound />} />
+        </Routes>
       </AuthProvider>
     </Router>
   </div>
