@@ -1,21 +1,21 @@
 import React, { useState } from 'react';
 import { Container, Card, Button, Alert } from 'react-bootstrap';
 import { useAuth } from '../context/AuthContext';
-import { Link, useHistory } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
 
 import { ROUTES } from '../config/constants.js';
 
 export default function Dashboard() {
   const [error, setError] = useState('');
   const { currentUser, logout } = useAuth();
-  const history = useHistory();
+  const navigate = useNavigate();
 
   async function handleLogout() {
     setError('');
 
     try {
       await logout();
-      history.push('/login');
+      navigate('/login');
     } catch {
       setError('Failed to log out');
     }
